Add more repository tests and remove sorts
