+++ tests/Aspire.Dashboard.Tests/TelemetryRepositoryTests/TraceTests.cs
@@ -1,0 +1,294 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using Aspire.Dashboard.Otlp.Model;
+using Aspire.Dashboard.Otlp.Storage;
+using Google.Protobuf.Collections;
+using OpenTelemetry.Proto.Trace.V1;
+using Xunit;
+using static Aspire.Dashboard.Tests.TelemetryRepositoryTests.TestHelpers;
+
+namespace Aspire.Dashboard.Tests.TelemetryRepositoryTests;
+
+public class TraceTests
+{
+    private static readonly DateTime s_testTime = new(1970, 1, 1, 0, 0, 0, DateTimeKind.Utc);
+
+    [Fact]
+    public void AddTraces()
+    {
+        // Arrange
+        var repository = CreateRepository();
+
+        // Act
+        var addContext = new AddContext();
+        repository.AddTraces(addContext, new RepeatedField<ResourceSpans>()
+        {
+            new ResourceSpans
+            {
+                Resource = CreateResource(),
+                ScopeSpans =
+                {
+                    new ScopeSpans
+                    {
+                        Scope = CreateScope(),
+                        Spans =
+                        {
+                            CreateSpan(traceId: "1", spanId: "1-1", startTime: s_testTime.AddMinutes(1), endTime: s_testTime.AddMinutes(10)),
+                            CreateSpan(traceId: "1", spanId: "1-2", startTime: s_testTime.AddMinutes(5), endTime: s_testTime.AddMinutes(10), parentSpanId: "1-1")
+                        }
+                    }
+                }
+            }
+        });
+
+        // Assert
+        Assert.Equal(0, addContext.FailureCount);
+
+        var applications = repository.GetApplications();
+        Assert.Collection(applications,
+            app =>
+            {
+                Assert.Equal("TestService", app.ApplicationName);
+                Assert.Equal("TestId", app.InstanceId);
+            });
+
+        var traces = repository.GetTraces(new GetTracesContext
+        {
+            ApplicationServiceId = applications[0].InstanceId,
+            StartIndex = 0,
+            Count = 10
+        });
+        Assert.Collection(traces.PagedResult.Items,
+            trace =>
+            {
+                AssertId("1", trace.TraceId);
+                AssertId("1-1", trace.FirstSpan.SpanId);
+                AssertId("1-1", trace.RootSpan!.SpanId);
+                Assert.Equal(2, trace.Spans.Count);
+            });
+    }
+
+    [Fact]
+    public void AddTraces_Traces_MultipleOutOrOrder()
+    {
+        // Arrange
+        var repository = CreateRepository();
+
+        // Act
+        var addContext1 = new AddContext();
+        repository.AddTraces(addContext1, new RepeatedField<ResourceSpans>()
+        {
+            new ResourceSpans
+            {
+                Resource = CreateResource(),
+                ScopeSpans =
+                {
+                    new ScopeSpans
+                    {
+                        Scope = CreateScope(),
+                        Spans =
+                        {
+                            CreateSpan(traceId: "1", spanId: "1-2", startTime: s_testTime.AddMinutes(5), endTime: s_testTime.AddMinutes(10), parentSpanId: "1-1")
+                        }
+                    }
+                }
+            }
+        });
+
+        var addContext2 = new AddContext();
+        repository.AddTraces(addContext2, new RepeatedField<ResourceSpans>()
+        {
+            new ResourceSpans
+            {
+                Resource = CreateResource(),
+                ScopeSpans =
+                {
+                    new ScopeSpans
+                    {
+                        Scope = CreateScope(),
+                        Spans =
+                        {
+                            CreateSpan(traceId: "2", spanId: "2-1", startTime: s_testTime.AddMinutes(3), endTime: s_testTime.AddMinutes(10))
+                        }
+                    }
+                }
+            }
+        });
+
+        var applications = repository.GetApplications();
+        Assert.Collection(applications,
+            app =>
+            {
+                Assert.Equal("TestService", app.ApplicationName);
+                Assert.Equal("TestId", app.InstanceId);
+            });
+
+        var traces1 = repository.GetTraces(new GetTracesContext
+        {
+            ApplicationServiceId = applications[0].InstanceId,
+            StartIndex = 0,
+            Count = 10
+        });
+        Assert.Collection(traces1.PagedResult.Items,
+            trace =>
+            {
+                AssertId("2", trace.TraceId);
+                AssertId("2-1", trace.FirstSpan.SpanId);
+                AssertId("2-1", trace.RootSpan!.SpanId);
+            },
+            trace =>
+            {
+                AssertId("1", trace.TraceId);
+                AssertId("1-2", trace.FirstSpan.SpanId);
+                Assert.Null(trace.RootSpan);
+            });
+
+        var addContext3 = new AddContext();
+        repository.AddTraces(addContext3, new RepeatedField<ResourceSpans>()
+        {
+            new ResourceSpans
+            {
+                Resource = CreateResource(),
+                ScopeSpans =
+                {
+                    new ScopeSpans
+                    {
+                        Scope = CreateScope(),
+                        Spans =
+                        {
+                            CreateSpan(traceId: "1", spanId: "1-1", startTime: s_testTime.AddMinutes(1), endTime: s_testTime.AddMinutes(10))
+                        }
+                    }
+                }
+            }
+        });
+
+        var traces2 = repository.GetTraces(new GetTracesContext
+        {
+            ApplicationServiceId = applications[0].InstanceId,
+            StartIndex = 0,
+            Count = 10
+        });
+        Assert.Collection(traces2.PagedResult.Items,
+            trace =>
+            {
+                AssertId("1", trace.TraceId);
+                AssertId("1-1", trace.FirstSpan.SpanId);
+                AssertId("1-1", trace.RootSpan!.SpanId);
+            },
+            trace =>
+            {
+                AssertId("2", trace.TraceId);
+                AssertId("2-1", trace.FirstSpan.SpanId);
+                AssertId("2-1", trace.RootSpan!.SpanId);
+            });
+    }
+
+    [Fact]
+    public void AddTraces_Spans_MultipleOutOrOrder()
+    {
+        // Arrange
+        var repository = CreateRepository();
+
+        // Act
+        repository.AddTraces(new AddContext(), new RepeatedField<ResourceSpans>()
+        {
+            new ResourceSpans
+            {
+                Resource = CreateResource(),
+                ScopeSpans =
+                {
+                    new ScopeSpans
+                    {
+                        Scope = CreateScope(),
+                        Spans =
+                        {
+                            CreateSpan(traceId: "1", spanId: "1-2", startTime: s_testTime.AddMinutes(2), endTime: s_testTime.AddMinutes(10), parentSpanId: "1-1"),
+                            CreateSpan(traceId: "1", spanId: "1-5", startTime: s_testTime.AddMinutes(5), endTime: s_testTime.AddMinutes(10), parentSpanId: "1-1"),
+                            CreateSpan(traceId: "1", spanId: "1-3", startTime: s_testTime.AddMinutes(3), endTime: s_testTime.AddMinutes(10), parentSpanId: "1-1"),
+                            CreateSpan(traceId: "1", spanId: "1-4", startTime: s_testTime.AddMinutes(4), endTime: s_testTime.AddMinutes(10), parentSpanId: "1-1"),
+                            CreateSpan(traceId: "1", spanId: "1-1", startTime: s_testTime.AddMinutes(1), endTime: s_testTime.AddMinutes(10))
+                        }
+                    }
+                }
+            }
+        });
+
+        var traces = repository.GetTraces(new GetTracesContext
+        {
+            ApplicationServiceId = null,
+            StartIndex = 0,
+            Count = 10
+        });
+        Assert.Collection(traces.PagedResult.Items,
+            trace =>
+            {
+                AssertId("1", trace.TraceId);
+                AssertId("1-1", trace.FirstSpan.SpanId);
+                AssertId("1-1", trace.RootSpan!.SpanId);
+                Assert.Collection(trace.Spans,
+                    s => AssertId("1-1", s.SpanId),
+                    s => AssertId("1-2", s.SpanId),
+                    s => AssertId("1-3", s.SpanId),
+                    s => AssertId("1-4", s.SpanId),
+                    s => AssertId("1-5", s.SpanId));
+            });
+    }
+
+    [Fact]
+    public void GetTraces_ReturnCopies()
+    {
+        // Arrange
+        var repository = CreateRepository();
+
+        // Act
+        var addContext1 = new AddContext();
+        repository.AddTraces(addContext1, new RepeatedField<ResourceSpans>()
+        {
+            new ResourceSpans
+            {
+                Resource = CreateResource(),
+                ScopeSpans =
+                {
+                    new ScopeSpans
+                    {
+                        Scope = CreateScope(),
+                        Spans =
+                        {
+                            CreateSpan(traceId: "1", spanId: "1-1", startTime: s_testTime.AddMinutes(5), endTime: s_testTime.AddMinutes(10))
+                        }
+                    }
+                }
+            }
+        });
+
+        var traces1 = repository.GetTraces(new GetTracesContext
+        {
+            ApplicationServiceId = null,
+            StartIndex = 0,
+            Count = 10
+        });
+        Assert.Collection(traces1.PagedResult.Items,
+            trace =>
+            {
+                AssertId("1", trace.TraceId);
+                AssertId("1-1", trace.FirstSpan.SpanId);
+                AssertId("1-1", trace.RootSpan!.SpanId);
+            });
+
+        var traces2 = repository.GetTraces(new GetTracesContext
+        {
+            ApplicationServiceId = null,
+            StartIndex = 0,
+            Count = 10
+        });
+        Assert.NotSame(traces1.PagedResult.Items[0], traces2.PagedResult.Items[0]);
+        Assert.NotSame(traces1.PagedResult.Items[0].Spans[0].Trace, traces2.PagedResult.Items[0].Spans[0].Trace);
+
+        var trace1 = repository.GetTrace(GetHexId("1"))!;
+        var trace2 = repository.GetTrace(GetHexId("1"))!;
+        Assert.NotSame(trace1, trace2);
+        Assert.NotSame(trace1.Spans[0].Trace, trace2.Spans[0].Trace);
+    }
+}
